REFACTOR/ [2차] network_api_manager.dart 함수 만듦 userThisHeaders만 사용하기 바람 range 있는 경우에는 (range: 값) 으로 사용하시면 됩니다
